zoneminder: Patch for NetBSD/earmv7hf-el
Patch more types, mostly time_t which is not long.

Do not assume linux if arm.  Rather than using a linux syscall to
determine if neon is present, just assume it is, optimizing for RPI2/3
and hoping little harm is done, as zm on a RPI1 is a stretch anyway
due to lack of RAM.

+++ zoneminder/patches/patch-src_zmu.cpp
@@ -16,7 +16,7 @@ time types
          } else {
            if ( have_output ) printf( "%c", separator );
 -          printf( "%ld.%02ld", timestamp.tv_sec, timestamp.tv_usec/10000 );
-+          printf( "%ld.%02ld", timestamp.tv_sec, (long) timestamp.tv_usec/10000 );
++          printf( "%jd.%02ld", (intmax_t) timestamp.tv_sec, (long) timestamp.tv_usec/10000 );
            have_output = true;
          }
        }
